Add missing page titles and move User component to /dashboard page
